adminhome deleting selected user tweets working with ajax and api

+++ public/js/useApi.js
@@ -6,13 +6,15 @@
  * Clears the table data
  */
 function clearTable() {
-  document.getElementById('.tablebody').innerHTML = '';
+  document.getElementById('tablebody').innerHTML = '';
 }
 
 function populateTable(tweetList) {
   for (var i = 0; i < tweetList.length; i++) {
     $('.tablebody ').append('<tr><td><b>' + tweetList[i].name + '</b>  </td><td>'
-        + tweetList[i].message + '</td><td class=\"delCheck\" class=\"ui checkbox\"><input name="+_id+" type=\"checkbox\"><label>Delete</label></div></td></tr>');
+        + tweetList[i].message + '</td><td ><input class=\"delCheck\"  value="'+tweetList[i]._id+'" type=\"checkbox\"><label>Delete</label></div></td></tr>');
+
+    console.log(tweetList[i]._id);
   }
 }
 
@@ -42,16 +44,62 @@ function searchUserTweets(){
   };
 
 function delSelectedTweets(){
-  var ids =[];
-  ids = $('.delCheck').val();
-  console.log(ids);
+  var id ;
+  //id = $('.delCheck:checked').val();
+  //id = document.querySelector('.delCheck:checked').value;
+  id = document.getElementsByClassName('delCheck');
+  //id = $('#checkbox').is('checked');
+  console.log("id content message "+id);
   console.log("do i get here")
-  /*for (let i =0; i < ids.length; i++){
+
+    for (let i = 0; i < id.length; i++) {
+      if (id[i].checked) {
+        console.log ("should be true "+ id[i].checked);
+        console.log("should be an id "+ id[i].value);
+        $.ajax({
+          dataType: 'json',
+          url: 'http://lap-austin:4000/api/tweets/' + id[i].value,
+          type: "DELETE",
+
+          success: function (data) {
+            console.log('success');
+            console.log(data);
+
+          },
+
+          error: function (err) {
+            console.log('fail');
+            console.log(err.statusText);
+            //updateResult(userName + ' ' + err.statusText);
+        },
+      });
+    }
+  }
+
+  //
+  var tweetList = [];
+  console.log("there should still be id : " + id);
+  for (let i = 0; i < id.length; i++) {
+    console.log("this should be false " + id[i].checked)
     $.ajax({
       dataType: 'json',
-      url: 'http://lap-austin:4000/api/tweet/'+ids[i],
-      type : "DELETE",
-    })
-  }*/
+      url: 'http://lap-austin:4000/api/tweet/' + id[i].value,
+      type: "GET",
+
+      success(data){
+        clearTable();
+        tweetList.push(data);
+        console.log('success');
+        console.log('tweetlist : ' + tweetList);
+        populateTable(tweetList);
+        console.log(data);
+      },
+      error: function (err) {
+        console.log(err.statusText);
+      }
+
+    });
+  }
+
 }
